Lab 6/requirements.txt
Django>=5.2.3
djangorestframework>=3.16.0
Pillow>=11.2.1
requests>=2.32.4
